+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## unreleased
+
+* [CHANGE] TSDB: Fix the predicate checking for blocks which are beyond the retention period to include the ones right at the retention boundary. #9633
+
 ## 2.50.1 / 2024-02-26
 
 * [BUGFIX] API: Fix metadata API using wrong field names. #13633